Address Daniil's test feedback
