+++ lldb/test/API/commands/expression/ptrauth/TestPtrAuthExpressions.py
@@ -4,14 +4,15 @@ from lldbsuite.test.lldbtest import *
 from lldbsuite.test import lldbutil
 
 
-class TestPtrAuthFixups(TestBase):
+class TestPtrAuthExpressions(TestBase):
     NO_DEBUG_INFO_TESTCASE = True
 
     @skipUnlessArm64eSupported
     def test_static_function_pointer(self):
         """On arm64e, function pointers are automatically signed (PAC).
-        Test that we can call a function through a function pointer from the
-        expression evaluator, which requires "fixing up" the pointer signing."""
+        Test that we can call a function through a static function pointer
+        from the expression evaluator, which requires "fixing up" the pointer
+        signing via the InjectPointerSigningFixups pass."""
         self.build()
 
         lldbutil.run_to_source_breakpoint(
@@ -31,11 +32,11 @@ class TestPtrAuthFixups(TestBase):
         )
 
     @skipUnlessArm64eSupported
-    def test_local_function_pointer(self):
-        """Test that function pointers with automatic (scoped) storage duration
-        work correctly. These exercise the PointerAuthCalls codegen path where
-        pointers are implicitly signed via codegen, rather than through the
-        InjectPointerSigningFixups pass."""
+    def test_indirect_call_through_caller(self):
+        """Test that a function pointer passed to a debuggee function is
+        correctly signed. The caller() function in the debuggee forces a
+        genuine indirect call, preventing the compiler from folding the
+        function pointer call into a direct call."""
         self.build()
 
         lldbutil.run_to_source_breakpoint(
@@ -43,13 +44,13 @@ class TestPtrAuthFixups(TestBase):
         )
 
         self.expect_expr(
-            "int (*fp)(int, int) = &add; fp(2, 3);",
+            "caller(add, 2, 3);",
             result_type="int",
             result_value="5",
         )
 
         self.expect_expr(
-            "int (*fp)(int, int) = &mul; fp(3, 7);",
+            "caller(mul, 3, 7);",
             result_type="int",
             result_value="21",
         )
@@ -58,8 +59,9 @@ class TestPtrAuthFixups(TestBase):
     def test_debuggee_signed_pointer(self):
         """Test that a signed function pointer stored in the debuggee's memory
         can be read and called from a user expression. The global_fp variable
-        in the debuggee holds a pointer signed with the debuggee's keys; since
-        expressions execute in the debuggee's process, auth should succeed."""
+        is signed with the IB key (__ptrauth(1, 0, 0)), which is
+        process-specific; this verifies that auth succeeds because expressions
+        execute in the debuggee's process, not the debugger's."""
         self.build()
 
         lldbutil.run_to_source_breakpoint(
